Move other 2 functions to controllers
